add ps:wait command and supporting function
This adds a function `ps_wait` that is mostly pulled from
uhoh-itsmaciek/heroku-ps-wait. This allows a user to run `heroku
ps:wait` to wait for all dynos to be up and running. Additionally, this
adds a flag to `ps:restart` (`--wait`) to run the `ps_wait` function
after initializing the restart.